% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_fit.R
\name{rain_fit}
\alias{rain_fit}
\title{A Function to fit models on rain data}
\usage{
rain_fit(y, X, raindata)
}
\arguments{
\item{y}{response}

\item{X}{covariate}

\item{raindata}{input data}
}
\value{
Summary of fitted models
}
\description{
rain_fit fits a linear model and a gamma glm to the rainfall
             dataset inputed by the user
}
