% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_plot.R
\name{rain_plot}
\alias{rain_plot}
\title{Function that produces plots}
\usage{
rain_plot(y, obj, rain_data)
}
\arguments{
\item{y}{Response}

\item{obj}{Object defined as the output of the model fit function}

\item{rain_data}{input for data}
}
\value{
A combined plot of the estimates from each model and the actual  #'         data
}
\description{
Function that produces plots
}
