% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_summary.R
\name{rain_summary}
\alias{rain_summary}
\title{Produce summary statistics for the rain data}
\usage{
rain_summary(y, X, obj)
}
\arguments{
\item{y}{Response}

\item{X}{covariate}

\item{obj}{Input obj that contains models run}
}
\value{
Summary statistics
}
\description{
rain_summary
}
